fix: paritioning failed on simple mode with UEFI
Constraints created from ped_constraint_exact are
too restrict, we can leave some room for libparted
so it can make better choice for us.

In this particular case
https://github.com/linuxdeepin/internal-discussion/issues/810
, GPT requires the last 33 sectors left for special
use(see https://superuser.com/a/1310579), if we
strictly declare we want that part, libparted will
throw "Unable to satisfy all the constraints on
the parition" error and failed to create the partition.

Change-Id: Iaf28a16825447ce747afb0b7873e0660679d26bc